refactor: split `index.ts` in `util`
