+++ apps/calendar/src/components/template.tsx
@@ -5,7 +5,7 @@ import isString from 'tui-code-snippet/type/isString';
 import { useStore } from '@src/contexts/calendarStore';
 import { templateSelector } from '@src/selectors';
 import { TemplateName } from '@src/template/default';
-import { identity } from '@src/util';
+import { identity } from '@src/util/utils';
 
 interface Props {
   template: TemplateName;
