+++ apps/calendar/src/controller/base.ts
@@ -3,8 +3,8 @@ import EventModel from '@src/model/eventModel';
 import EventUIModel from '@src/model/eventUIModel';
 import TZDate from '@src/time/date';
 import { makeDateRange, MS_PER_DAY, toEndOfDay, toFormat, toStartOfDay } from '@src/time/datetime';
-import { isSameEvent } from '@src/util';
 import Collection from '@src/util/collection';
+import { isSameEvent } from '@src/util/events';
 
 export type IDS_OF_DAY = Record<string, number[]>;
